requirements.txt
pydicom>=2.3.0
numpy>=1.20.0
pillow>=9.0.0
pylibjpeg>=2.0.0
pylibjpeg-libjpeg>=2.1.0 